Refactoring: profile state
